Update admin and streamer pages
- Fixed object assignment reactivity

- Added more "show key" toggles (these will be nicer and less intrusive with
better styling down the line)

- Made helper functions for fetch calls (There is some duplication still between
the two pages. I have a utils.ts WIP that will solve this later but I have to
properly implement a context object to pass the necessary values through the
function calls. Maybe there's completely different better way to do this. I'm
not going to rush it)

- Pressing enter on initial login for both pages now works

- Probably another fix or two I can't remember :)